Added basic commands, also command for compliling. Performed first test.
I was able to compile simple blink program. Transcription from LabVIEW to arduino is quite easy when dealing with VI it's straightforward, one VI = one arduino cmd. But there has to be parsing for structures, so easy way how to do is, put in the beginning of structure small VI which scan it and include begin and end of structure.

Possible problems with structures parsing:
  - figure out what is wired to the end terminal, ie. how to transcript end of structure (while loop)